build(deamdify): invoke deamdify through browserify global switch rather than in jquery-ui package.j
To bundle jquery-ui with browserify we have to convert jquery-ui's AMD bundling to CommonJS.
deamidfy does this for us.  we had been invoking deamidfy by manually inserting the deamdify trigger
into the jquery-ui package.json (that npm installed in node_modules), but this manual change can't
be made when using Travis.  We now instead invoke deamidfy as part of the browserify invokation in
our package.json, so manual changes are no longer needed.