feat(tag note): load type att dynamically and allow for custom input
Type options are pulled from the schema. If the schema doesn't define the options, the default
options will be displayed. If the schema allows, the user can insert a custom type, or have no type
at all. The interface was changed from radiobox to dropdown to improove ux. A popup window shows up
when space is hit while typing a custom type warning about this type of action on xml att.

BREAKING CHANGE: options are pulled dinamically. Checks need to be done while building the dialog.

#274