feat: 🎸 Tagging major refactoring
WIP: Add the ability to use tags already present in the document to
filter and limit the options to further tagging. The position of the tag
in the document respects the schema rules. The file got reorganized and
usses class keyword (es6 sintatic sugar). Optimized for Orlando Entry
Schema. Supoprt to other schemas is comming.