feat: 🎸 Transform CWRC-Writer into a React Project
Introduce a base react layer on the root of the project to which other
components will attached on. Introduce overmindJS to manage apps state.
This is just the first step to convert the whole project into react.

BREAKING CHANGE: 🧨 CWRC-Writer-base-React behaviours less as a function to attach to the
window and more like a react compnent. Change configuration and
signature methods. More break-changes to come.