fix: 🐛 cwrc2xml conversion: selfclosing tags only if no children or text node child with empty textcontent
previous version test for content on the tag itself. If the tag
contained selfclosing chdild tag (no content), conversion would
self-close the current tag removing the child in the process. This issue
affect a range of tags such as BIBCIT, REG, AREA, ITEM, and TEXTSCOPE on
ORLANDO schema to name a few. TEI Schema might had affected as well. The
current fix test the contetn of the child of the tag instead the tag
itself, keeping the tag open if a child is self-closed.

✅ Closes: #300